backend/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-filter==2.4.0
djoser==2.1.0
drf-extra-fields==3.1.1
pillow==8.3.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-debug-toolbar==3.2.2
reportlab==3.6.9